blueprint testing
