+++ src/project/profile/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+profile_bp = Blueprint('profile', __name__, template_folder='../templates')
+
+from . import routes
