+++ src/project/templates/sign_up.html
@@ -14,7 +14,7 @@
 <body>
 	<nav>
 	<ul>
-		<li><a href="/login">Login</a></li>
+		<li><a href="/login">Login!!!!</a></li>
 		<li><a href="/register">Create account</a></li>
 	</ul>
 	</nav>
